feat: moved simulation button inside action builder
